Yep, I did that massive thing again...
It all went out of hand when I decided to strip the calculation of pressure loss from V1, and have the users supply pressure loss in the model.

There are a few tests that still fail. But I also have way too many test data files that are [almost] the same. It should be cleaned up to avoid confusion.